// types.ts
export type UserStatus = 'Activo' | 'Inactivo' | 'Suspendido' | 'Egresado' | 'Pendiente';
// ...
export interface SavedView {
    id: string;
    name: string;
    filters: any;
}